Updated tests to use fn directly
