+++ src/app.js
@@ -1,5 +1,6 @@
 // Main module file
 require('dotenv').config();
+const {filterQueryParams} = require('./utils/query-params');
 
 const fastify = require('fastify')({
     logger: {
@@ -29,26 +30,8 @@ fastify.register(require('@fastify/http-proxy'), {
     rewritePrefix: '', // we'll hardcode this in PROXY_TARGET
     httpMethods: ['GET', 'POST', 'PUT', 'DELETE'],
     rewriteRequest: (req) => {
-        // Get query parameters without creating a full URL object
-        const searchParams = new URLSearchParams(req.url.split('?')[1] || '');
-        
-        // Extract only token and name
-        const token = searchParams.get('token');
-        const name = searchParams.get('name');
-        
-        // Create new query string with only these parameters
-        const newSearchParams = new URLSearchParams();
-        if (token) {
-            newSearchParams.set('token', token);
-        }
-        if (name) {
-            newSearchParams.set('name', name);
-        }
-        
-        // Update the request URL (keep pathname, replace query)
-        const path = req.url.split('?')[0];
-        req.url = path + (newSearchParams.toString() ? `?${newSearchParams.toString()}` : '');
-        
+        // Filter query parameters to only include token and name
+        req.url = filterQueryParams(req.url);
         return req;
     },
     replyOptions: {
